factored some stuff out
