+++ bin/invtool
@@ -1,26 +1,6 @@
 #!/usr/bin/env python2.7
 import sys
-import argparse
-from invtool.options import *
-from invtool.intr_options import *
-from invtool.dispatch import build_dns_parsers, dispatch
+from invtool.main import main
 
 if __name__ == "__main__":
-    inv_parser = argparse.ArgumentParser(prog='invtool')
-    format_group = inv_parser.add_mutually_exclusive_group()
-    format_group.add_argument('--json', default=False, dest='p_json',
-                            help="Format the output as JSON",
-                            action='store_true')
-    format_group.add_argument('--silent', default=False, dest='p_silent',
-                            help="Silence all stdout and stderr",
-                            action='store_true')
-    format_group.add_argument('--debug', default=False, dest='DEBUG',
-                            help="Print stuff",
-                            action='store_true')
-    base_parser = inv_parser.add_subparsers(dest='dtype')
-    build_dns_parsers(base_parser)
-    nas = inv_parser.parse_args(sys.argv[1:])
-    resp_code, resp_list = dispatch(nas)
-    if not nas.p_silent and resp_list:
-        print '\n'.join(resp_list),
-    sys.exit(resp_code)
+    sys.exit(main(sys.argv))
