+++ invtool/lib/registrar.py
@@ -1,0 +1,8 @@
+class Registrar():
+    dns_dispatches = []
+    dispatches = []
+
+    def register(self, dispatch):
+        self.dispatches.append(dispatch)
+
+registrar = Registrar()
