--- invtool/test_parse.py
@@ -1,123 +1,0 @@
-from invtool import *
-
-args = [
-        '--format json A create --label foo --domain bar.com --ip 10.2.3.4 --no-private --no-public',
-        '--format text A create --label foo --domain bar.com --ip 10.2.3.4 --ttl 3434 --no-public',
-        '--format json A update --pk 1 --domain bar.com --ip 10.2.3.4',
-        '--format text A update --pk 1 --label foo --domain bar.com --ip 10.2.3.4 --no-private',
-        '--format json A delete --pk 1',
-        '--format text A delete --pk 4',
-        'A create --label foo --domain bar.com --ip 10.2.3.4 --ttl 3434 --no-private --no-public',
-        'A create --label foo --domain bar.com --ip 10.2.3.4',
-        'A update --pk 1 --label foo --domain bar.com',
-        'A update --pk 1 --label foo --ip 10.2.3.4',
-        'A delete --pk 1',
-        'A delete --pk 4',
-        '--format json AAAA create --label foo --domain bar.com --ip 10.2.3.4 --no-private',
-        '--format text AAAA create --label foo --domain bar.com --ip 10.2.3.4 --no-public',
-        '--format json AAAA update --pk 1  --ip 10.2.3.4 --ttl 3434',
-        '--format text AAAA update --pk 1 --label foo --domain bar.com --ip 10.2.3.4',
-        '--format json AAAA delete --pk 4',
-        '--format text AAAA delete --pk 54', 'AAAA create --label foo --domain bar.com --ip 10.2.3.4 --no-private',
-        'AAAA create --label foo --domain bar.com --ip 10.2.3.4',
-        'AAAA update --pk 1 --label foo --ip 10.2.3.4',
-        'AAAA update --pk 1 --label foo --domain bar.com --ip 10.2.3.4',
-        'AAAA delete --pk 3',
-        'AAAA delete --pk 5',
-
-        '--format json PTR create --target foo.bar.com --ip 10.2.3.4',
-        '--format text PTR create --target foo.bar.com --ip 10.2.3.4 --ttl 3434 --no-private --no-public',
-        '--format json PTR update --pk 1 --ttl 3434',
-        '--format text PTR update --pk 1 --target foo.bar.com --ip 10.2.3.4',
-        '--format json PTR delete --pk 345',
-        '--format text PTR delete --pk 345',
-        'PTR create --target foo.bar.com --ip 10.2.3.4 --no-public',
-        'PTR create --target foo.bar.com --ip 10.2.3.4',
-        'PTR update --pk 1 --target foo.bar.com',
-        'PTR update --pk 1 --target foo.bar.com --ip 10.2.3.4',
-        'PTR delete --pk 3',
-        'PTR delete --pk 5',
-
-        '--format json CNAME create --target foo.bar.com --label foo --domain baz.com',
-        '--format text CNAME create --target foo.bar.com --label foo --domain baz.com --ttl 3434 --no-private',
-        '--format json CNAME update --pk 1',
-        '--format text CNAME update --label foo --domain baz.com --pk 3',
-        '--format json CNAME delete --pk 345',
-        '--format text CNAME delete --pk 345',
-        'CNAME create --target foo.bar.com --label foo --domain baz.com',
-        'CNAME create --target foo.bar.com --label foo --domain baz.com',
-        'CNAME update --pk 1 --target foo.bar.com --label foo --domain baz.com --no-public',
-        'CNAME update --pk 1 --target foo.bar.com --label foo --domain baz.com',
-        'CNAME delete --pk 3',
-        'CNAME delete --pk 5',
-
-        '--format json NS create --target foo.bar.com --domain baz.com',
-        '--format text NS create --target foo.bar.com --domain baz.com --ttl 3434',
-        '--format json NS update --pk 1 --target foo.bar.com  --domain baz.com --ttl 3434',
-        '--format text NS update --target foo.bar.com  --domain baz.com --pk 3',
-        '--format json NS delete --pk 345',
-        '--format text NS delete --pk 345',
-        'NS create --target foo.bar.com  --domain baz.com',
-        'NS create --target foo.bar.com  --domain baz.com',
-        'NS update --pk 1 --target foo.bar.com  --domain baz.com',
-        'NS update --pk 1 --target foo.bar.com  --domain baz.com',
-        'NS delete --pk 3',
-        'NS delete --pk 5',
-
-        '--format json MX create --target foo.bar.com --label foo --domain baz.com --priority 3',
-        '--format text MX create --target foo.bar.com --label foo --domain baz.com --ttl 3434 --priority 3',
-        '--format json MX update --pk 1 --target foo.bar.com --label foo --domain baz.com --ttl 3434',
-        '--format text MX update --target foo.bar.com --label foo --domain baz.com --pk 3 --priority 3',
-        '--format json MX delete --pk 345',
-        '--format text MX delete --pk 345',
-        'MX create --target foo.bar.com --label foo --domain baz.com --priority 3',
-        'MX create --target foo.bar.com --label foo --domain baz.com --priority 3',
-        'MX update --pk 1 --target foo.bar.com --label foo --domain baz.com --priority 3',
-        'MX update --pk 1 --target foo.bar.com --label foo --priority 33 --domain baz.com',
-        'MX delete --pk 3',
-        'MX delete --pk 5',
-
-        '--format json SRV create --target foo.bar.com --label _foo --domain baz.com --priority 3 --weight 34 --port 44',
-        '--format text SRV create --target foo.bar.com --label _foo --domain baz.com --ttl 3434 --priority 3 --weight 34 --port 44',
-        '--format json SRV update --pk 1 --target foo.bar.com --label _foo --domain baz.com --ttl 3434 --priority 3 --weight 34 --port 44',
-        '--format text SRV update --target foo.bar.com --label foo --domain baz.com --pk 3 --priority 3 --weight 34 --port 44',
-        '--format json SRV delete --pk 4',
-        '--format text SRV delete --pk 5',
-        'SRV create --target foo.bar.com --label _foo --domain baz.com --priority 3 --weight 34 --port 44',
-        'SRV create --target foo.bar.com --label _foo --domain baz.com --priority 3 --weight 34 --port 44 --weight 34 --port 44',
-        'SRV update --pk 1 --target foo.bar.com --label _foo --domain baz.com --priority 3',
-        'SRV update --pk 1 --target foo.bar.com --label _foo --priority 33 --domain baz.com --weight 34 --port 44',
-        'SRV delete --pk 3',
-        'SRV delete --pk 5',
-
-        '--format json TXT create --text "foo.bar_is_cool" --label foo --domain baz.com',
-        '--format text TXT create --text "foo.bar_is_cool" --label foo --domain baz.com --ttl 3434',
-        '--format json TXT update --pk 1 --text "foo.bar_is_cool" --label foo --domain baz.com --ttl 3434',
-        '--format text TXT update --text "foo.bar_is_cool" --label foo --domain baz.com --pk 3',
-        '--format json TXT delete --pk 345',
-        '--format text TXT delete --pk 345',
-        'TXT create --text "foo.bar_is_cool" --label foo --domain baz.com',
-        'TXT create --text "foo.bar_is_cool" --label foo --domain baz.com',
-        'TXT update --pk 1 --text "foo.bar_is_cool" --label foo --domain baz.com',
-        'TXT update --pk 1 --text "foo.bar_is_cool" --label foo --domain baz.com',
-        'TXT delete --pk 3',
-        'TXT delete --pk 5',
-        'SSHFP create --label foo --domain baz.com --algo RSA --finger-type SHA1 --key foo',
-
-        '--format json INTR create --label foo --domain bar.com --ip 10.2.3.4 --no-private --iname eth0',
-        '--format text INTR create --label foo --domain bar.com --ip 10.2.3.4 --no-public --iname eth0.2 --no-dns --no-dhcp',
-        '--format json INTR update --pk 1  --ip 10.2.3.4 --ttl 3434',
-        '--format text INTR update --pk 1 --label foo --domain bar.com --ip 10.2.3.4',
-        '--format json INTR delete --pk 4',
-        '--format text INTR delete --pk 54', 'INTR create --label foo --domain bar.com --ip 10.2.3.4 --no-private',
-        'INTR create --label foo --domain bar.com --ip 10.2.3.4',
-        'INTR update --pk 1 --label foo --ip 10.2.3.4 --no-dhcp',
-        'INTR update --pk 1 --label foo --domain bar.com --ip 10.2.3.4 --no-dns',
-        'INTR delete --pk 3',
-        'INTR delete --pk 5',
-    ]
-
-for arg in args:
-    print '-' * 20
-    print arg
-    print dispatch_dns(inv_parser.parse_args(arg.split()))
